+++ src/Sync/IOriginSync.php
@@ -2,7 +2,6 @@
 
 namespace srag\Plugins\Hub2\Sync;
 
-use Exception;
 use srag\Plugins\Hub2\Exception\AbortOriginSyncException;
 use srag\Plugins\Hub2\Exception\AbortSyncException;
 use srag\Plugins\Hub2\Exception\BuildObjectsFailedException;
@@ -29,14 +28,6 @@ interface IOriginSync {
 	 * @throws AbortSyncException
 	 */
 	public function execute();
-
-
-	/**
-	 * Get a collection of all exceptions occurred during executing the sync.
-	 *
-	 * @return Exception[]
-	 */
-	public function getExceptions();
 
 
 	/**