+++ src/Sync/Summary/IOriginSyncSummary.php
@@ -25,14 +25,6 @@ interface IOriginSyncSummary {
 
 
 	/**
-	 * @param IOriginSync $originSync
-	 *
-	 * @return string
-	 */
-	public function getSummaryOfOrigin(IOriginSync $originSync);
-
-
-	/**
 	 * @return void
 	 */
 	public function sendNotifications();