+++ src/Sync/Summary/OriginSyncSummaryMail.php
@@ -16,22 +16,7 @@ class OriginSyncSummaryMail extends OriginSyncSummaryBase implements IOriginSync
 	/**
 	 * @inheritdoc
 	 */
-	public function getOutputAsString() {
-		$return = "";
-		foreach ($this->syncs as $sync) {
-			$return .= $this->renderOneSync($sync) . "\n\n";
-		}
-
-		return $return;
-	}
-
-
-	/**
-	 * @param IOriginSync $originSync
-	 *
-	 * @return string
-	 */
-	private function renderOneSync(IOriginSync $originSync) {
+	protected function renderOneSync(IOriginSync $originSync): string {
 		// Print out some useful statistics: --> Should maybe be a OriginSyncSummary object
 		$msg = self::plugin()->translate("summary_for", "", [ $originSync->getOrigin()->getTitle() ]) . "\n**********\n";
 		$msg .= self::plugin()->translate("summary_delivered_data_sets", "", [ $originSync->getCountDelivered() ]) . "\n";
@@ -40,27 +25,13 @@ class OriginSyncSummaryMail extends OriginSyncSummaryBase implements IOriginSync
 		$msg .= self::plugin()->translate("summary_updated", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_UPDATED) ]) . "\n";
 		$msg .= self::plugin()->translate("summary_outdated", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_OUTDATED) ]) . "\n";
 		$msg .= self::plugin()->translate("summary_ignored", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_IGNORED) ]) . "\n";
-		foreach ($originSync->getNotifications()->getMessages() as $context => $messages) {
-			$msg .= "$context: \n**********\n";
-			foreach ($messages as $message) {
-				$msg .= "$message\n";
-			}
-			$msg .= "\n";
-		}
-		$msg .= self::plugin()->translate("summary_exceptions") . "\n**********\n";
-		foreach ($originSync->getExceptions() as $exception) {
+
+		$msg .= self::plugin()->translate("summary_logs") . "\n**********\n";
+		foreach ($originSync->getLogs() as $exception) {
 			$msg .= $exception->getMessage() . "\n";
 		}
 		$msg = rtrim($msg, "\n");
 
 		return $msg;
-	}
-
-
-	/**
-	 * @inheritdoc
-	 */
-	public function getSummaryOfOrigin(IOriginSync $originSync) {
-		return $this->renderOneSync($originSync);
 	}
 }