+++ src/Sync/Summary/OriginSyncSummaryWeb.php
@@ -16,22 +16,7 @@ class OriginSyncSummaryWeb extends OriginSyncSummaryBase implements IOriginSyncS
 	/**
 	 * @inheritdoc
 	 */
-	public function getOutputAsString() {
-		$return = "";
-		foreach ($this->syncs as $sync) {
-			$return .= $this->renderOneSync($sync) . "\n\n";
-		}
-
-		return $return;
-	}
-
-
-	/**
-	 * @param IOriginSync $originSync
-	 *
-	 * @return string
-	 */
-	private function renderOneSync(IOriginSync $originSync) {
+	protected function renderOneSync(IOriginSync $originSync): string {
 		// Print out some useful statistics: --> Should maybe be a OriginSyncSummary object
 		$msg = self::plugin()->translate("summary_for", "", [ $originSync->getOrigin()->getTitle() ]) . "\n**********\n";
 		$msg .= self::plugin()->translate("summary_delivered_data_sets", "", [ $originSync->getCountDelivered() ]) . "\n";
@@ -40,33 +25,19 @@ class OriginSyncSummaryWeb extends OriginSyncSummaryBase implements IOriginSyncS
 		$msg .= self::plugin()->translate("summary_updated", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_UPDATED) ]) . "\n";
 		$msg .= self::plugin()->translate("summary_outdated", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_OUTDATED) ]) . "\n";
 		$msg .= self::plugin()->translate("summary_ignored", "", [ $originSync->getCountProcessedByStatus(IObject::STATUS_IGNORED) ]) . "\n";
-		foreach ($originSync->getNotifications()->getMessages() as $context => $messages) {
-			$msg .= "$context: \n**********\n";
-			foreach ($messages as $message) {
-				$msg .= "$message\n";
-			}
-			$msg .= "\n";
-		}
-		$msg .= self::plugin()->translate("summary_exceptions") . "\n**********\n";
-		if (count($originSync->getExceptions()) > 0) {
+
+		$msg .= self::plugin()->translate("summary_logs") . "\n**********\n";
+		if (count($originSync->getLogs()) > 0) {
 			// Only output one exception on web because it's to large for Session (ilUtil::sendInfo)
-			$exception = current($originSync->getExceptions());
+			$exception = current($originSync->getLogs());
 			$msg .= $exception->getMessage() . "\n";
 
-			if (count($originSync->getExceptions()) > 1) {
-				$msg .= self::plugin()->translate("summary_exceptions_more", "", [ (count($originSync->getExceptions()) - 1) ]) . "\n";
+			if (count($originSync->getLogs()) > 1) {
+				$msg .= self::plugin()->translate("summary_logs_more", "", [ (count($originSync->getLogs()) - 1) ]) . "\n";
 			}
 		}
 		$msg = rtrim($msg, "\n");
 
 		return $msg;
-	}
-
-
-	/**
-	 * @inheritdoc
-	 */
-	public function getSummaryOfOrigin(IOriginSync $originSync) {
-		return $this->renderOneSync($originSync);
 	}
 }